requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-telegram-bot==21.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.5
urllib3==2.2.1
